+++ src/main/webapp/static/js/script-home.js
@@ -18,23 +18,6 @@ document.addEventListener('DOMContentLoaded', function () {
         });
     }
 
-    // 회원가입 버튼 이벤트 리스너 추가
-    const joinBtn = document.getElementById('joinBtn');
-    const mobileJoinBtn = document.getElementById('mobileJoinBtn');
-
-    if (joinBtn) {
-        joinBtn.addEventListener('click', function () {
-            document.getElementById('joinModal').style.display = 'flex';
-        });
-    }
-
-    if (mobileJoinBtn) {
-        mobileJoinBtn.addEventListener('click', function () {
-            document.getElementById('mobileMenuModal').style.display = 'none';
-            document.getElementById('joinModal').style.display = 'flex';
-        });
-    }
-
     // 모달 닫기 이벤트 리스너
     document.addEventListener('click', function (e) {
         if (e.target.classList.contains('close')) {
